Build Model and Deconvolute in one step

Build Model and Deconvolute in one step

Usage

build_and_deconvolute(
  single_cell_obj,
  spatial_obj,
  method = NULL,
  cell_type_col = "cell_ontology_class",
  batch_id_col = NULL,
  assay_sc = "counts",
  assay_sp = "counts",
  return_object = TRUE,
  verbose = FALSE,
  ...
)

Arguments

single_cell_obj

Single Cell Object containing reference data to build the model

spatial_obj

SpatialExperiment to be deconvoluted

method

deconvolution method

cell_type_col

column of single_cell_obj containing cell type information

batch_id_col

column of SpatialObject containing batch_id information

assay_sc

the assay of the single cell object to use, default = "counts"

assay_sp

the assay of the spatialExperiment to use, default = "counts"

return_object

if true return anotation spatialExperiment, if false return table

verbose

output more information

...

further parameters passed to the selected function