Deconvolution with spacedeconv

Perform cell type deconvolution with spacedeconv.

Usage

deconvolute(
  spatial_obj,
  signature = NULL,
  single_cell_obj = NULL,
  cell_type_col = "cell_ontology_class",
  method = NULL,
  batch_id_col = NULL,
  assay_sc = "counts",
  assay_sp = "counts",
  return_object = TRUE,
  verbose = FALSE,
  ...
)

Arguments

spatial_obj

A SpatialExperiment

signature

Gene Expression Signature

single_cell_obj

A SingleCellExperiment

cell_type_col

Column name of the single_cell_obj where the cell type can be found

method

Deconvolution Method to use, see deconvolution_methods() for a full list of available methods

batch_id_col

column name of batch ids in single cell object

assay_sc

which single cell assay to use for computation

assay_sp

which spatial assay to use for computation

return_object

Return an Object or result Table, TRUE = Object

verbose

display more detailed information

...

Further parameters passed to the selected deconvolution method

Value

The deconvolution result as a table

Examples

# more examples can be found in the documentation website
data("spatial_data_2")

spatial_data_2 <- spacedeconv::preprocess(spatial_data_2)
#> ── spacedeconv ─────────────────────────────────────────────────────────────────
#> ℹ testing parameter
#> Loading required package: SpatialExperiment
#> Loading required package: SingleCellExperiment
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following object is masked from ‘package:spacedeconv’:
#> 
#>     normalize
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> ✔ parameter OK [1.4s]
#> 
#> ℹ Removing 85 observations with umi count below threshold
#> ✔ Removed 85 observations with umi count below threshold [433ms]
#> 
#> ℹ Removing 13418 variables with all zero expression
#> ✔ Removed 13418 variables with all zero expression [14ms]
#> 
spatial_data_2 <- spacedeconv::normalize(spatial_data_2)
#> ── spacedeconv ─────────────────────────────────────────────────────────────────
#> ℹ testing parameter
#> ✔ parameter OK [18ms]
#> 
#> ℹ Normalizing using cpm
#> ✔ Finished normalization using cpm [2s]
#> 
#> ℹ Please note the normalization is stored in an additional assay

deconvolution <- spacedeconv::deconvolute(
  spatial_obj = spatial_data_2,
  method = "estimate",
)
#> ── spacedeconv ─────────────────────────────────────────────────────────────────
#> ℹ testing parameter
#> ✔ parameter OK [26ms]
#> 
#> 
#> 
#> ── Spatial 
#> Assays: "counts" and "cpm"
#> Genes: 23173
#> → without expression: 0 (0%)
#> Spots: 1225
#> Spots under tissue: 1225 (100%)
#> Median Genes Per Spot: 4795
#> → without expression: 0 (0%)
#> Umi count range: 536 - 64683
#> Spots with UMI count below 500: 0 (0%)
#> ✔ Rownames set
#> ✔ Colnames set
#> ℹ deconvoluting
#> Warning: replacing previous import ‘utils::findMatches’ by ‘S4Vectors::findMatches’ when loading ‘AnnotationDbi’
#> 
#> >>> Running estimate
#> [1] "Merged dataset includes 9317 genes (1095 mismatched)."
#> [1] "1 gene set: StromalSignature  overlap= 135"
#> [1] "2 gene set: ImmuneSignature  overlap= 140"
#> ✔ finished [10.9s]
#>