Comparative Scatterplot of Two SpatialExperiments

Creates a scatterplot to compare deconvolution results between two SpatialExperiment objects.

Usage

plot_scatter(
  spe = NULL,
  value1,
  value2,
  spe1 = NULL,
  spe2 = NULL,
  log_scale = FALSE,
  dot_color = "#1f77b4",
  point_alpha = 0.8,
  fix_coords = FALSE,
  coord_range = NULL,
  title = "Comparative Scatterplot"
)

Arguments

spe

SpatialExperiment with quantification results

value1

Deconvolution result to plot from the first object.

value2

Deconvolution result to plot from the second object.

spe1

First SpatialExperiment object.

spe2

Second SpatialExperiment object.

log_scale

Logical, whether to use log scale for axes.

dot_color

Color for the dots in the plot.

point_alpha

Alpha for the dots, controlling transparency.

fix_coords

Logical, whether to fix coordinate system to be equal.

coord_range

Numeric vector of length 2 to set coordinate limits.

title

Title of the plot.

Value

A scatterplot